Makes a document as interface
